Update protocols to proper header/C++ format. Move to separate dir
